pyproject.toml
[tool.poetry]
name = "fastapi-security"
version = "0.1.0"
description = ""
authors = ["elcollins"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
sqlmodel = "^0.0.16"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
